Refactor pokepaste.tsx and use-poke-paste.tsx for better code organization and functionality
